Added Testing While Refactoring for Multiple Platforms
This is the command-complete draft of the refactoring to support for CentOS and Ubuntu.

The workflow walks through adding this for the install recipe. A lab asks the learner to work through the service recipe. It finishes with adding support for the new platform in Test Kitchen and running an Integration Test.