Added the Why Write Tests Section
This is a discussion based section that attempts to engage all the learners present in the training. Specifically this is to bring in the individuals that may feel outside of the group; those not convinced there is value in testing. This section does not try to disuade them and instead it does my best work to try and embrace them and invite them to listen and discuss their feelings.