Update README headers and fix bad link
Cannot decide on Abstract or Description so I'll take both. I also like Environments more instead of what was there before. While shaving the yak on titles I found a real bug. The link out to the recipe that 'creates' the AMI for this training was not pointing to the correct one.