Update 03 Writing a Test Speaker Notes for first 10 slides
There are some speaker notes and some initial slides to those notes implemented. More visualizations would help but for right now this was the longest series of notes at the beginning of the course. When I teach this in a few days I will evaluate them and make some more changes.