Update 08 Refactoring Multiple Platforms is now conceptual slide
Way more yak than ever before. This yak has a spiritual side.
Includes the acknowledgement and lab associated with the configuration recipe. This creates a good conversation as the learner will refactor and then undo the change and talk about it.
Made the first experience with pry a group activity